[Do Not Merge] reference PR for buggy media sdk
This is a reference PR for an attempt to upgrade the libMediaSDK to
version 3.0.1.2, which was published in April 2025.

This version of the media SDK, or at least the Debian installer
libMediaSDK-dev-3.0.1.2-20250418_153137-amd64.deb, is buggy. I got the
following error when trying to execute `MediaSDKTest --help` from the
container shell.

```bash
root@51e1f959aae2:~# MediaSDKTest --help
glfwDisplayMgr: failed to setup GLFW!
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'
GLFW error: '65537', 'The GLFW library is not initialized'

*** ORYOL ASSERT: window != nullptr
  msg=none
  file=/media/nature/workspace1/InsMediaSDKGPU/deps/bmg/Source/arvoryol/code/Modules/Gfx/private/glfw/glfwDisplayMgr.cc
  line=48
  func=Oryol::_priv::OffscreenWindowPool::InitPool(size_t)::<lambda()>
  callstack:
/lib/libMediaSDK.so(_ZN5Oryol10StackTrace4DumpEPci+0x3d) [0x79ba6c4ab82d]
/lib/libMediaSDK.so(_ZN5Oryol3Log9AssertMsgEPKcS2_S2_iS2_+0x104) [0x79ba6c4a9b44]
/lib/libMediaSDK.so(+0x30601de) [0x79ba6c4cd1de]
/lib/x86_64-linux-gnu/libc.so.6(+0x99ee8) [0x79ba68ec2ee8]
/lib/libMediaSDK.so(_ZN5Oryol5_priv19OffscreenWindowPool8InitPoolEm+0x6d) [0x79ba6c4cc2bd]
/lib/x86_64-linux-gnu/libc.so.6(+0x99ee8) [0x79ba68ec2ee8]
/lib/libMediaSDK.so(_ZN5Oryol5_priv14glfwDisplayMgr4InitEm+0x6d) [0x79ba6c4cc6fd]
/lib/libMediaSDK.so(_ZN5Oryol19PlatformSetupStage1EPv+0x14) [0x79ba6c4cd6a4]
/lib/libMediaSDK.so(_ZN3ins7InitEnvEv+0x60) [0x79ba69fbc570]
MediaSDKTest(+0x86291) [0x5db58435f291]
/lib/x86_64-linux-gnu/libc.so.6(+0x29d90) [0x79ba68e52d90]
/lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0x80) [0x79ba68e52e40]
MediaSDKTest(+0xa9ef5) [0x5db584382ef5]

Aborted (core dumped)

```